+++ includes/class.CommonFunctions.inc.php
@@ -131,7 +131,7 @@ class CommonFunctions
     public static function executeProgram($strProgramname, $strArgs, &$strBuffer, $booErrorRep = true)
     {
         if (defined('PSI_LOG') && is_string(PSI_LOG) && (strlen(PSI_LOG)>0) && ((substr(PSI_LOG, 0, 1)=="-") || (substr(PSI_LOG, 0, 1)=="+"))) {
-            $out = self::_parse_log_file("Executing: ".$strProgramname.' '.$strArgs);
+            $out = self::_parse_log_file("Executing: ".trim($strProgramname.' '.$strArgs));
             if ($out == false) {
                 if (substr(PSI_LOG, 0, 1)=="-") {
                     $strBuffer = '';
@@ -144,7 +144,7 @@ class CommonFunctions
                 return true;
             }
         }
-        
+
         $strBuffer = '';
         $strError = '';
         $pipes = array();
@@ -175,13 +175,8 @@ class CommonFunctions
             $process = proc_open($strProgram." ".$strArgs, $descriptorspec, $pipes);
         }
         if (is_resource($process)) {
-            if (defined("PSI_MODE_POPEN") && PSI_MODE_POPEN === true) {
-                $pipes[0] = null;
-                $pipes[2] = fopen("/dev/null", "r");
-            }
             self::_timeoutfgets($pipes, $strBuffer, $strError);
             if (defined("PSI_MODE_POPEN") && PSI_MODE_POPEN === true) {
-                fclose($pipes[2]);
                 $return_value = pclose($pipes[1]);
             } else {
                 fclose($pipes[0]);
@@ -201,7 +196,7 @@ class CommonFunctions
         $strError = trim($strError);
         $strBuffer = trim($strBuffer);
         if (defined('PSI_LOG') && is_string(PSI_LOG) && (strlen(PSI_LOG)>0) && (substr(PSI_LOG, 0, 1)!="-") && (substr(PSI_LOG, 0, 1)!="+")) {
-            error_log("---".gmdate('r T')."--- Executing: ".$strProgramname.' '.$strArgs."\n".$strBuffer."\n", 3, PSI_LOG);
+            error_log("---".gmdate('r T')."--- Executing: ".trim($strProgramname.' '.$strArgs)."\n".$strBuffer."\n", 3, PSI_LOG);
         }
         if (! empty($strError)) {
             if ($booErrorRep) {
@@ -241,7 +236,7 @@ class CommonFunctions
                 return true;
             }
         }
-        
+
         $strFile = "";
         $intCurLine = 1;
         $error = Error::singleton();
@@ -335,7 +330,7 @@ class CommonFunctions
      */
     public static function checkForExtensions($arrExt = array())
     {
-        if ((PSI_OS == "Minix") || (PSI_SYSTEM_CODEPAGE == "UTF-8"))
+        if ((strcasecmp(PSI_SYSTEM_CODEPAGE,"UTF-8") == 0) || (strcasecmp(PSI_SYSTEM_CODEPAGE,"CP437") == 0))
             $arrReq = array('simplexml', 'pcre', 'xml', 'dom');
         elseif (PSI_OS == "WINNT")
             $arrReq = array('simplexml', 'pcre', 'xml', 'mbstring', 'dom', 'com_dotnet');
@@ -369,21 +364,46 @@ class CommonFunctions
      * @param array   $pipes array of file pointers for stdin, stdout, stderr (proc_open())
      * @param string  &$out  target string for the output message (reference)
      * @param string  &$err  target string for the error message (reference)
-     * @param integer $sek   timeout value in seconds
+     * @param integer $timeout   timeout value in seconds (default value is 30)
      *
      * @return void
      */
-    private static function _timeoutfgets($pipes, &$out, &$err, $sek = 30)
+    private static function _timeoutfgets($pipes, &$out, &$err, $timeout = 30)
     {
-        // fill output string
-        $time = $sek;
-        $w = null;
-        $e = null;
+        $w = NULL;
+        $e = NULL;
 
-        $read = array($pipes[1],$pipes[2]);
-        while (!(feof($pipes[1]) && feof($pipes[2])) && ($n = stream_select($read, $w, $e, $time)) !== false && $n > 0) {
-                $out .= fread($pipes[1], 4096);
-                $err .= fread($pipes[2], 4096);
+        if (defined("PSI_MODE_POPEN") && PSI_MODE_POPEN === true) {
+            $pipe2 = false;  
+        } else {
+            $pipe2 = true; 
+        }
+        while (!(feof($pipes[1]) || ($pipe2 && feof($pipes[2])))) {
+            if ($pipe2) {
+                $read = array($pipes[1], $pipes[2]);
+            } else {
+                $read = array($pipes[1]);
+            }
+
+            $n = stream_select($read, $w, $e, $timeout);
+
+            if ($n === FALSE) {
+                error_log('stream_select: failed !');
+                break;
+            }
+            else if ($n === 0) {
+                error_log('stream_select: timeout expired !');
+                break;
+            }
+
+            foreach ($read as $r) {
+                if ($r == $pipes[1]) {
+                    $out .= fread($r, 4096);
+                }
+                if ($pipe2 && ($r == $pipes[2])) {
+                    $err .= fread($r, 4096);
+                }
+            }
         }
     }