+++ js/phpSysInfo/phpsysinfo.js
@@ -1015,10 +1015,10 @@ function refreshTemp(xml) {
         var label = "", value = "", limit = 0, _limit = "";
         label = $(this).attr("Label");
         value = $(this).attr("Value").replace(/\+/g, "");
-        limit = parseFloat($(this).attr("Max").replace(/\+/g, ""));
+        limit = ($(this).attr("Max") !== undefined) ? parseFloat($(this).attr("Max").replace(/\+/g, "")) : 'NaN';
         if (isFinite(limit))
             _limit = formatTemp(limit, xml);
-        $("#tempTable").append("<tr><td>" + label + "</td><td class=\"right\">" + formatTemp(value, xml) + "</td><td class=\"right\">" + _limit + "</td></tr>");
+        $("#tempTable tbody").append("<tr><td>" + label + "</td><td class=\"right\">" + formatTemp(value, xml) + "</td><td class=\"right\">" + _limit + "</td></tr>");
         values = true;
     });
     if (values) {
@@ -1087,6 +1087,32 @@ function refreshFan(xml) {
 }
 
 /**
+ * (re)fill the power block with the values from the given xml<br><br>
+ * build the power information into a separate block, if there is no power information available the
+ * entire table will be removed to avoid HTML warnings
+ * @param {jQuery} xml phpSysInfo-XML
+ */
+function refreshPower(xml) {
+    var values = false;
+    $("#powerTable tbody").empty();
+    $("MBInfo Power Item", xml).each(function getPowers(id) {
+        var label = "", value = "", limit = 0, _limit = "";
+        label = $(this).attr("Label");
+        value = $(this).attr("Value").replace(/\+/g, "");
+        limit = ($(this).attr("Max") !== undefined) ? parseFloat($(this).attr("Max").replace(/\+/g, "")) : 'NaN';
+        if (isFinite(limit))
+            _limit = round(limit, 2) + "&nbsp;" + genlang(103, true);
+        $("#powerTable tbody").append("<tr><td>" + label + "</td><td class=\"right\">" + round(value, 2) + "&nbsp;" + genlang(103, true) + "</td><td class=\"right\">" + _limit + "</td></tr>");
+        values = true;
+    });
+    if (values) {
+        $("#power").show();
+    }
+    else {
+        $("#power").remove();
+    }
+}
+/**
  * (re)fill the ups block with the values from the given xml<br><br>
  * build the ups information into a separate block, if there is no ups information available the
  * entire table will be removed to avoid HTML warnings
@@ -1101,7 +1127,7 @@ function refreshUps(xml) {
 
     $("#ups").empty();
     $("UPSInfo UPS", xml).each(function getUps(id) {
-        var name = "", model = "", mode = "", start_time = "", upsstatus = "", temperature = "", outages_count = "", last_outage = "", last_outage_finish = "", line_voltage = "", load_percent = "", battery_voltage = "", battery_charge_percent = "", time_left_minutes = "";
+        var name = "", model = "", mode = "", start_time = "", upsstatus = "", temperature = "", outages_count = "", last_outage = "", last_outage_finish = "", line_voltage = "", load_percent = "", battery_date = "", battery_voltage = "", battery_charge_percent = "", time_left_minutes = "";
         name = $(this).attr("Name");
         model = $(this).attr("Model");
         mode = $(this).attr("Mode");
@@ -1114,6 +1140,7 @@ function refreshUps(xml) {
         last_outage_finish = $(this).attr("LastOutageFinish");
         line_voltage = $(this).attr("LineVoltage");
         load_percent = parseInt($(this).attr("LoadPercent"), 10);
+        battery_date = $(this).attr("BatteryDate");
         battery_voltage = $(this).attr("BatteryVoltage");
         battery_charge_percent = parseInt($(this).attr("BatteryChargePercent"), 10);
         time_left_minutes = $(this).attr("TimeLeftMinutes");
@@ -1148,6 +1175,10 @@ function refreshUps(xml) {
         }
         if (!isNaN(load_percent)) {
             html += "<tr><td style=\"width:160px\">" + genlang(78, false) + "</td><td>" + createBar(load_percent) + "</td></tr>\n";
+            tree.push(index);
+        }
+        if (battery_date !== undefined) {
+            html += "<tr><td style=\"width:160px\">" + genlang(104, false) + "</td><td>" + battery_date + "</td></tr>\n";
             tree.push(index);
         }
         if (battery_voltage !== undefined) {
@@ -1214,6 +1245,7 @@ function reload() {
             refreshVoltage(xml);
             refreshFan(xml);
             refreshTemp(xml);
+            refreshPower(xml);
             refreshUps(xml);
 
             $('.stripeMe tr:nth-child(even)').addClass('even');
@@ -1266,6 +1298,7 @@ $(document).ready(function buildpage() {
             refreshTemp(xml);
             refreshVoltage(xml);
             refreshFan(xml);
+            refreshPower(xml);
             refreshUps(xml);
 
             changeLanguage();