+++ CHANGELOG.md
@@ -1,9 +1,32 @@
+Changelog of phpSysInfo
+=======================
+
+http://phpsysinfo.sourceforge.net/
+
+phpSysInfo 3.1.7
+----------------
+
+ - [ADD] Ksplice support for Linux
+ - [ADD] Show CPU frequency max and min for Darwin (Mac OS X)
+ - [ADD] Show System Language and Code Page on Darwin (Mac OS X)
+ - [ADD] Show network interfaces infos for Minix and SunOS
+ - [ADD] SMS, gNewSense and Vector to detected distros
+ - [ADD] LMSensors power information
+ - [ADD] Battery installation date for the UPS info
+
+ - [UPD] Network interfaces infos and filesystems infos for FreeBSD
+ - [UPD] Updated support of SunOS
+ - [UPD] Memory informations on Darwin systems
+ - [UPD] BAT plugin - updated Linux support
+ - [UPD] Updated HWSensors - OpenBSD sensor program
+
 phpSysInfo 3.1.6
+----------------
 
  - [ADD] Porteus, Peppermint, Manjaro, Netrunner and Salix to detected distros
  - [ADD] Show CPU frequency max for WINNT
  - [ADD] Show network interfaces infos for Darwin (Mac OS X)
-  
+
  - [UPD] SNMPPInfo plugin, ink level for some of the data
  - [UPD] jQuery 2.0.3 and jQuery 1.10.2
  - [UPD] Russian Translation ru.xml
@@ -12,6 +35,7 @@ phpSysInfo 3.1.6
  - [SEC] Fix JSONP
 
 phpSysInfo 3.1.5
+----------------
 
  - [ADD] Possibility to define multiple SENSOR_PROGRAM
  - [ADD] Added display of temperature and fan speed for IPMI sensor program
@@ -22,6 +46,7 @@ phpSysInfo 3.1.5
  - [FIX] Fix recovery detection of RAID arrays on debian systems #18
 
 phpSysInfo 3.1.4
+----------------
 
  - [ADD] Option for reading the results of functions executeProgram() and rfts() from log
  - [ADD] Show CPU frequency max and min for variable speed processors for Linux and Android
@@ -31,6 +56,7 @@ phpSysInfo 3.1.4
  - [UPD] jQuery 2.0.2 and jQuery 1.10.1
 
 phpSysInfo 3.1.3
+----------------
 
  - [ADD] IPFire, Sabayon, PearOS, ClearOS, Frugalware, Fuduntu, Foresight, Tinycore, ALT Linux, ROSA and RedHatEnterpriseServer to detected distros
 
@@ -41,6 +67,7 @@ phpSysInfo 3.1.3
  - [FIX] Run php-cs-fixer on php files (PSR-2 fixer)
 
 phpSysInfo 3.1.2
+----------------
 
  - [ADD] Tempsensor and CPU frequency for Raspberry Pi (thanks to hawkeyexp)
  - [ADD] Linaro to detected distros
@@ -50,6 +77,7 @@ phpSysInfo 3.1.2
  - [FIX] Incorrect display of chunk size for the plugin mdstatus for some results
 
 phpSysInfo 3.1.1
+----------------
 
  - [ADD] SolusOS, Deepin and antiX to detected distros
  - [ADD] Simplified Chinese translation
@@ -57,6 +85,7 @@ phpSysInfo 3.1.1
  - [UPD] jQuery 1.9.1
 
 phpSysInfo 3.1.0
+----------------
 
  - [NEW] Configuration moved from config.php and subdirs of "plugins" to one file phpsysinfo.ini